requirements.txt
flask==3.1.2
flask-mail==0.9.1
apscheduler==3.10.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
reportlab==4.4.4
werkzeug==3.1.3
blinker==1.9.0
click==8.3.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
pillow==12.0.0
charset-normalizer==3.4.4
colorama==0.4.6
